+++ src/Spotify.Mvc/Controllers/NacionalidadController.cs
@@ -1,9 +1,3 @@
-using Microsoft.AspNetCore.Mvc;
-using Microsoft.AspNetCore.Authorization;
-using Spotify.ReposDapper;
-using Spotify.Core.Entidades;
-using Spotify.Mvc.Models;
-
 namespace Spotify.Mvc.Controllers;
 
 [Authorize(Roles = "3")]
@@ -78,7 +72,7 @@ public class NationalityController : Controller
 
             var model = new NacionalidadModel
             {
-                IdNacionalidad = nacionalidad.IdNacionalidad,
+                IdNacionalidad = (int)nacionalidad.IdNacionalidad,
                 Pais = nacionalidad.Pais
             };
